AI-StT Align Configurate/requirements.txt
openai-whisper
ffmpeg-python
moviepy